requirements.txt
flask[async]~=2.2.2
flask_cors~=3.0.10
ansi2html~=1.8.0
ipython~=8.12.0
langchain~=0.1.13
openai~=1.11.1
chromadb~=0.4.2
streamlit~=1.24.1
loguru~=0.5.3
atlassian-python-api~=3.39.0
python-decouple~=3.8
tiktoken~=0.5.2
waitress~=2.1.2
gunicorn~=21.2.0
beautifulsoup4~=4.12.3
lxml~=5.1.0
appdirs~=1.4.4
markdownify~=0.11.6
flask_htmx~=0.3.2
flask_sqlalchemy~=3.1.1
boto3==1.34.71
fuzzywuzzy~=0.18.0
python-Levenshtein~=0.25.0
flask_login~=0.6.3
langchain-openai~=0.1.0
dotenv~=0.0.6